% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_ceiling.R
\name{decimal_ceiling}
\alias{decimal_ceiling}
\title{Rounding of decimals}
\usage{
decimal_ceiling(x, digits = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The desired number of digits after the decimal point.}
}
\value{
A numeric vector.
}
\description{
Takes a numeric vector x and returns a numeric vector containing the smallest numerics with \code{digits} places after the decimal not less than the corresponding elements of x.
}
\examples{
decimal_ceiling(x = 10.12345, digits = 3)
}
