% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_floor.R
\name{decimal_floor}
\alias{decimal_floor}
\title{Rounding of decimals}
\usage{
decimal_floor(x, digits = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The desired number of digits after the decimal point.}
}
\value{
A numeric vector.
}
\description{
Takes a numeric vector x and returns a numeric vector containing the largest numerics with \code{digits} places after the decimal not greater than the corresponding elements of x.
}
\examples{
decimal_floor(x = 10.12345, digits = 3)
}
