% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid.R
\name{expand_grid}
\alias{expand_grid}
\title{Faster alternative to \code{expand.grid()}}
\usage{
expand_grid(v1, v2)
}
\arguments{
\item{v1}{A vector.}

\item{v2}{A vector.}
}
\value{
A dataframe of combinations of \code{v1} and \code{v2}.
}
\description{
Faster alternative to \code{expand.grid()}
}
\examples{
expand_grid(v1 = c(1, 2), v2 = c(1, 2, 3, 4))
}
