% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log10_minor_breaks.R
\name{log10_minor_breaks}
\alias{log10_minor_breaks}
\title{Add log10 minor breaks to a ggplot with ggplot2::scale_x/y_log10() activated.}
\usage{
log10_minor_breaks(...)
}
\arguments{
\item{...}{Added via a \code{+} in a \code{ggplot2} chain.}
}
\value{
A log10 plot with minor breaks.
}
\description{
Add log10 minor breaks to a ggplot with ggplot2::scale_x/y_log10() activated.
}
\examples{
ggplot2::ggplot() + ggplot2::scale_x_log10('x', minor_breaks = log10_minor_breaks())
}
