% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padded_data.R
\name{padded_data.frame}
\alias{padded_data.frame}
\title{Creates a dataframe from a list of vectors not necessarily of equal length. If vectors are of equal length, the dataframe is padded with NA values.}
\usage{
padded_data.frame(l, ...)
}
\arguments{
\item{l}{List of vectors not necessarily of equal length.}
}
\value{
A data frame with vectors in columns. Unequal length is resolved with NA values.
}
\description{
Creates a dataframe from a list of vectors not necessarily of equal length. If vectors are of equal length, the dataframe is padded with NA values.
}
\examples{
l <- list(x = c(1, 2, 3), y = c('a', 'b'))
padded_data.frame(l)
}
