% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_that.R
\name{time_that}
\alias{time_that}
\title{Returns the run time for a given expression.}
\usage{
time_that(expr, units = "secs")
}
\arguments{
\item{expr}{Expression to time, contained within brackets}

\item{units}{Units to compute the run time}
}
\value{
Any assignments and the runtime
}
\description{
Returns the run time for a given expression.
}
\examples{
time_that(sum(rnorm(150)), units = 'secs')
}
