% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait.R
\name{wait}
\alias{wait}
\title{Pauses execution of remaining code until the user clicks a button to continue.}
\usage{
wait()
}
\value{
A button for the user to click when they are ready to continue execution.
}
\description{
Pauses execution of remaining code until the user clicks a button to continue.
}
