% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_range.R
\name{zero_range}
\alias{zero_range}
\title{Check equality among all elements of a vector.}
\usage{
zero_range(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{tol}{Tolerance one is willing to accept.}
}
\value{
A logical.
}
\description{
Check equality among all elements of a vector.
}
\examples{
zero_range(rep(1, times = 5))
}
